core: commonwealth: Move PrettyJSONResponse to utils
