+++ core/libs/commonwealth/commonwealth/utils/apis.py
@@ -1,0 +1,17 @@
+import json
+from typing import Any
+
+from starlette.responses import Response as StarletteResponse
+
+
+class PrettyJSONResponse(StarletteResponse):
+    media_type = "application/json"
+
+    def render(self, content: Any) -> bytes:
+        return json.dumps(
+            content,
+            ensure_ascii=False,
+            allow_nan=False,
+            indent=2,
+            separators=(", ", ": "),
+        ).encode(self.charset)
