+++ core/services/wifi/setup.py
@@ -45,5 +45,6 @@ setuptools.setup(
         "uvicorn == 0.13.4",
         "aiofiles == 0.6.0",
         "fastapi-versioning == 0.9.0",
+        "commonwealth == 0.1.0",
     ],
 )
